GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - .idea/.gitignore
   - .idea/hamster-home.iml
   - .idea/git_toolbox_prj.xml
   - app/home/xui/components/XuiSummary.tsx
   - .idea/jsLibraryMappings.xml
   - .idea/prettier.xml

Your previous branch was: refs/heads/main

The sha for that commit was: cfdbb8563c5193e37fa806779b34136365318e0c

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch